Initial version of HZdbg
Argument to executable specifies the path of the process to be traced
print a welcome message for entering HZDbg
fork a process and start tracing it and pause it as soon as it begins
print when it is paused
Resume it when user enters 'c'. It should print some message letting you know it has resumed.
After it is done free it from the Zombie state